+++ src/api/index.ts
@@ -1,2 +1,3 @@
 export * from "./nodeContents"
 export * from "./transports"
+export * from "./objectstructure"