+++ src/utilities.ts
@@ -36,3 +36,14 @@ export function xmlArray<T>(xml: any, ...path: string[]): T[] {
 
   return []
 }
+
+const ok = Object.keys
+export const xmlRoot = (o: any) => o[ok(o)[0]]
+// extract XML attributes of a node from its JSON representation
+export const xmlNodeAttr = (n: any) =>
+  ok(n)
+    .filter(k => k.match(/^(?!@_xmlns)@_/))
+    .reduce((part: any, cur) => {
+      part[cur.replace(/^@_/, "")] = n[cur]
+      return part
+    }, {})