+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to the "vscode-abap-remote-fs" extension will be documented 
 
 Format based on [Keep a Changelog](http://keepachangelog.com/)
 
+## [0.2.5] - 2019-01-22
+
+### Added
+
+- object structure/metadata
+- logic to extract links from object definitions
+
 ## [0.2.4] - 2019-01-21
 
 ### Added